+++ src/main/java/com/example/bookshopwebapplication/entities/OrderItem2.java
@@ -15,11 +15,11 @@ public class OrderItem2 {
     private Long productId;
     private String productName;
     private String productImage;
-    private BigDecimal basePrice;
-    private BigDecimal discountPercent;
-    private BigDecimal price;
+    private Double basePrice;
+    private Double discountPercent;
+    private Double price;
     private Integer quantity;
-    private BigDecimal subtotal;
+    private Double subtotal;
     private Timestamp createdAt;
     private Timestamp updatedAt;
 }
